CSI_ENDPOINT can accept file path
This patch updates the GoCSI dependency in order to make it so
CSI_ENDPOINT can now accept a file path. Instead of requiring an
explicit Golang network address, the value of CSI_ENDPOINT can now be
set to a file path. If the file path is valid then the library will
treat the file path as an implied use of a UNIX socket file.