Updated Volume Identity & MntPath in Metadata
This patch updates the VolumeInfo.Id.Values map to use the key "id" as
the volume's name. Additionally, the VolumeInfo.Metadata.Values map will
have a key "targetpaths" with a comma-separated list of a volume's mount
paths if the volume is mounted on the host on which the plug-in is
running.

The Volumeinfo.Metadata.Values map can also use the key "name" to
specify a friendly name for a volume.